moved addonLoader to own file
