+++ server/gsm-server/server.py
@@ -1,21 +1,6 @@
 import json
-import glob
-import zipfile as zipf
+from addonLoader import *
 from socketserver import TCPServer, StreamRequestHandler
-
-def addon_loader():
-    gameLoads = dict()
-    dirName = './games'
-    zipList = glob.glob(dirName + '/*.zip')
-
-
-    for zipname in zipList:
-        addonFile = zipf.ZipFile(zipname)
-        manifestFileRaw = addonFile.read('manifest.json')
-        manifestFile = json.loads(manifestFileRaw)
-        gameLoads = gameLoads | manifestFile
-
-    return (gameLoads)
 
 class Handler(StreamRequestHandler):
     games = addon_loader()
